+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `/api/v1/groups/followed` API endpoint.
 - Added `/api/v1/timelines/group/{group_id}` API endpoint.
+- Added `use_context` parameter to `/api/v1/statuses/{status_id}/load_conversation` API method.
 
 ### Changed